+++ doc/conf.py
@@ -340,5 +340,11 @@ sphinx_gallery_conf = {
     'within_subsection_order': NumberOfCodeLinesSortKey,
     'find_mayavi_figures': find_mayavi_figures,
     'expected_failing_examples': ['../examples/no_output/plot_raise.py',
-                                  '../examples/no_output/plot_syntaxerror.py']
+                                  '../examples/no_output/plot_syntaxerror.py'],
+    'binder': {'org': 'sphinx-gallery',
+               'repo': 'sphinx-gallery.github.io',
+               'url': 'https://mybinder.org',
+               'branch': 'master',
+               'dependencies': './binder/requirements.txt'
+               }
 }